chore: updated tests
