+++ service/src/test/java/org/jboss/sbomer/service/test/integ/rest/TestableApiVersion.java
@@ -1,0 +1,19 @@
+package org.jboss.sbomer.service.test.integ.rest;
+
+import lombok.Getter;
+
+public enum TestableApiVersion {
+    V1ALPHA3("v1alpha3"), V1BETA1("v1beta1");
+
+    @Getter
+    String path;
+
+    TestableApiVersion(String path) {
+        this.path = path;
+    }
+
+    @Override
+    public String toString() {
+        return path;
+    }
+}
